+++ scripts/generate-env.js
@@ -9,6 +9,7 @@ const project_root = path.resolve(__dirname, '..');
 const file = path.join(project_root, 'example', 'src', 'env.ts');
 const content = `export const test = true;
 export const appKey = [''];
+export const appId = [''];
 export const PushInfo = { sendId: undefined, KeyId: undefined };
 export const accounts = [{ id: undefined, mm: undefined }];
 export const groups = [{ id: undefined, owner: undefined, name: undefined }];